Added OrderItem Entity

+++ Store1/Model/CustomerAdditionalDetail.cs
@@ -9,7 +9,7 @@ namespace Store1.Model
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int ID { get; set; }
         [ForeignKey("CustomerID")]
-        public long CustomerID { get; set; }
+        public int CustomerID { get; set; }
         public Customer Customer { get; set; }
         public string City {  get; set; }
         public string Country { get; set; }
